# test-constructor/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для конструктора тестов",
        "title": "Test Constructor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/manager/events/{id}/statistics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить статистику всех попыток по мероприятию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Статистика мероприятия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Фильтр по дополнительным тестам",
                        "name": "is_extra",
                        "in": "body",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статистика мероприятия",
                        "schema": {
                            "$ref": "#/definitions/dto.StatisticsResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Мероприятие не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/tests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список тестов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Получить тесты",
                "responses": {
                    "200": {
                        "description": "Список тестов",
                        "schema": {
                            "$ref": "#/definitions/dto.TestsListResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новый тест с вопросами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Создать тест",
                "parameters": [
                    {
                        "description": "Данные теста",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Тест создан",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTestResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/tests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает тест с вопросами по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Получить тест",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали теста",
                        "schema": {
                            "$ref": "#/definitions/dto.TestDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Тест не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет тест по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Удалить тест",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тест удален",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteTestResponse"
                        }
                    },
                    "404": {
                        "description": "Тест не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список всех пользователей с ролью \"intern\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Список стажёров",
                "responses": {
                    "200": {
                        "description": "Список стажёров",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUsersResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить статистику всех попыток пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Статистика пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статистика пользователя",
                        "schema": {
                            "$ref": "#/definitions/dto.UserStatisticsResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.ChoiceOption": {
            "type": "object",
            "properties": {
                "is_true": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "domain.MatchingPair": {
            "type": "object",
            "properties": {
                "left": {
                    "type": "string"
                },
                "right": {
                    "type": "string"
                }
            }
        },
        "domain.QuestionOptions": {
            "type": "object",
            "properties": {
                "case_sensitive": {
                    "type": "boolean"
                },
                "choice": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ChoiceOption"
                    }
                },
                "correct_input": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "matching": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.MatchingPair"
                    }
                },
                "sequence": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SequenceItem"
                    }
                }
            }
        },
        "domain.SequenceItem": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateQuestionRequest": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/domain.QuestionOptions"
                },
                "order_number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTestRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateQuestionRequest"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTestResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.DeleteTestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Тест удален"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetUsersResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserInfo"
                    }
                }
            }
        },
...
        "dto.QuestionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "options": {
                    "$ref": "#/definitions/domain.QuestionOptions"
                },
                "order_number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.QuestionStatInfo": {
            "type": "object",
            "properties": {
                "is_correct": {
                    "type": "boolean"
                },
                "max_points": {
                    "type": "integer"
                },
                "order_number": {
                    "type": "integer"
                },
                "points_earned": {
                    "type": "integer"
                },
                "question_type": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "dto.StatisticsResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserAttemptInfo"
                    }
                }
            }
        },
        "dto.TestDetailResponse": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "integer"
                },
                "creator_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuestionResponse"
                    }
                },
                "test_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.TestResponse": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "integer"
                },
                "creator_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.TestsListResponse": {
            "type": "object",
            "properties": {
                "tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TestResponse"
                    }
                }
            }
        },
...
        "dto.UserAttemptDetail": {
            "type": "object",
            "properties": {
                "attempt_id": {
                    "type": "integer"
                },
                "event_name": {
                    "type": "string"
                },
                "is_extra": {
                    "type": "boolean"
                },
                "max_score": {
                    "type": "integer"
                },
                "passed": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuestionStatInfo"
                    }
                },
                "score": {
                    "type": "integer"
                },
                "test_title": {
                    "type": "string"
                }
            }
        },
        "dto.UserAttemptInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "is_extra": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "max_score": {
                    "type": "integer"
                },
                "passed": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuestionStatInfo"
                    }
                },
                "score": {
                    "type": "integer"
                },
                "time_spent_minutes": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "dto.UserStatisticsResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserAttemptDetail"
                    }
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {ваш_токен}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}